apis/v1: Split ServiceMonitor types in their own file
