+++ pkg/apis/monitoring/v1/types.go
@@ -32,10 +32,6 @@ const (
 	PrometheusesKind  = "Prometheus"
 	PrometheusName    = "prometheuses"
 	PrometheusKindKey = "prometheus"
-
-	ServiceMonitorsKind   = "ServiceMonitor"
-	ServiceMonitorName    = "servicemonitors"
-	ServiceMonitorKindKey = "servicemonitor"
 )
 
 var resourceToKind = map[string]string{
@@ -1202,58 +1198,6 @@ type AlertmanagerEndpoints struct {
 	EnableHttp2 *bool `json:"enableHttp2,omitempty"`
 }
 
-// +genclient
-// +k8s:openapi-gen=true
-// +kubebuilder:resource:categories="prometheus-operator",shortName="smon"
-
-// ServiceMonitor defines monitoring for a set of services.
-type ServiceMonitor struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of desired Service selection for target discovery by
-	// Prometheus.
-	Spec ServiceMonitorSpec `json:"spec"`
-}
-
-// ServiceMonitorSpec contains specification parameters for a ServiceMonitor.
-// +k8s:openapi-gen=true
-type ServiceMonitorSpec struct {
-	// JobLabel selects the label from the associated Kubernetes service which will be used as the `job` label for all metrics.
-	//
-	// For example:
-	// If in `ServiceMonitor.spec.jobLabel: foo` and in `Service.metadata.labels.foo: bar`,
-	// then the `job="bar"` label is added to all metrics.
-	//
-	// If the value of this field is empty or if the label doesn't exist for the given Service, the `job` label of the metrics defaults to the name of the Kubernetes Service.
-	JobLabel string `json:"jobLabel,omitempty"`
-	// TargetLabels transfers labels from the Kubernetes `Service` onto the created metrics.
-	TargetLabels []string `json:"targetLabels,omitempty"`
-	// PodTargetLabels transfers labels on the Kubernetes `Pod` onto the created metrics.
-	PodTargetLabels []string `json:"podTargetLabels,omitempty"`
-	// A list of endpoints allowed as part of this ServiceMonitor.
-	Endpoints []Endpoint `json:"endpoints"`
-	// Selector to select Endpoints objects.
-	Selector metav1.LabelSelector `json:"selector"`
-	// Selector to select which namespaces the Kubernetes Endpoints objects are discovered from.
-	NamespaceSelector NamespaceSelector `json:"namespaceSelector,omitempty"`
-	// SampleLimit defines per-scrape limit on number of scraped samples that will be accepted.
-	SampleLimit uint64 `json:"sampleLimit,omitempty"`
-	// TargetLimit defines a limit on the number of scraped targets that will be accepted.
-	TargetLimit uint64 `json:"targetLimit,omitempty"`
-	// Per-scrape limit on number of labels that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelLimit uint64 `json:"labelLimit,omitempty"`
-	// Per-scrape limit on length of labels name that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelNameLengthLimit uint64 `json:"labelNameLengthLimit,omitempty"`
-	// Per-scrape limit on length of labels value that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelValueLengthLimit uint64 `json:"labelValueLengthLimit,omitempty"`
-	// Attaches node metadata to discovered targets.
-	// Requires Prometheus v2.37.0 and above.
-	AttachMetadata *AttachMetadata `json:"attachMetadata,omitempty"`
-}
-
 // Endpoint defines a scrapeable endpoint serving Prometheus metrics.
 // +k8s:openapi-gen=true
 type Endpoint struct {
@@ -1500,17 +1444,6 @@ func (c *TLSConfig) Validate() error {
 	return nil
 }
 
-// ServiceMonitorList is a list of ServiceMonitors.
-// +k8s:openapi-gen=true
-type ServiceMonitorList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard list metadata
-	// More info: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-	// List of ServiceMonitors
-	Items []*ServiceMonitor `json:"items"`
-}
-
 // MetadataConfig configures the sending of series metadata to the remote storage.
 // +k8s:openapi-gen=true
 type MetadataConfig struct {
@@ -1563,16 +1496,6 @@ func (l *Prometheus) DeepCopyObject() runtime.Object {
 
 // DeepCopyObject implements the runtime.Object interface.
 func (l *PrometheusList) DeepCopyObject() runtime.Object {
-	return l.DeepCopy()
-}
-
-// DeepCopyObject implements the runtime.Object interface.
-func (l *ServiceMonitor) DeepCopyObject() runtime.Object {
-	return l.DeepCopy()
-}
-
-// DeepCopyObject implements the runtime.Object interface.
-func (l *ServiceMonitorList) DeepCopyObject() runtime.Object {
 	return l.DeepCopy()
 }
